[infra] collapse dock_drive and driver, part of #31
