--- kobuki_dock_drive/.gitignore
@@ -1,3 +1,0 @@
-bin
-build
-lib
