--- kobuki_dock_drive/include/CMakeLists.txt
@@ -1,10 +1,0 @@
-##############################################################################
-# Headers
-##############################################################################
-
-install(
-  DIRECTORY
-    ${PROJECT_NAME}/
-  DESTINATION
-    include
-)
