--- kobuki_dock_drive/package.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0"?>
-<package format="2">
-  <name>kobuki_dock_drive</name>
-  <version>1.1.1</version>
-  <description>
-	    Dock driving library for Kobuki. Users owning a docking station for Kobuki
-	    can use this tool to develop autonomous docking drive algorithms.
-  </description>
-  <author>Daniel Stonier</author>
-  <author>Younghun Ju</author>
-  <maintainer email="d.stonier@gmail.com">Daniel Stonier</maintainer>
-  <license>BSD</license>
-  <url type="bugtracker">https://github.com/kobuki-base/kobuki_core/issues</url>
-  <url type="repository">https://github.com/kobuki-base/kobuki_core</url>
-  <url type="website">https://kobuki.readthedocs.io/en/release-1.0.x/</url>
-
-  <buildtool_depend>ament_cmake_ros</buildtool_depend>
-
-  <build_depend>ecl_build</build_depend>
-  <build_depend>ecl_threads</build_depend>
-  <build_depend>ecl_geometry</build_depend>
-  <build_depend>ecl_linear_algebra</build_depend>
-
-  <exec_depend>ecl_threads</exec_depend>
-  <exec_depend>ecl_geometry</exec_depend>
-  <exec_depend>ecl_linear_algebra</exec_depend>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
