--- kobuki_dock_drive/src/CMakeLists.txt
@@ -1,39 +1,0 @@
-##############################################################################
-# Sources
-##############################################################################
-
-file(GLOB SOURCES RELATIVE ${CMAKE_CURRENT_SOURCE_DIR} *.cpp)
-
-##############################################################################
-# Libraries
-##############################################################################
-
-add_library(${PROJECT_NAME} ${SOURCES})
-
-target_include_directories(
-  ${PROJECT_NAME}
-  PUBLIC
-    $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/include>
-    $<INSTALL_INTERFACE:include>
-)
-
-target_link_libraries(
-  ${PROJECT_NAME}
-  PUBLIC
-    ecl_geometry::ecl_geometry
-    ecl_linear_algebra::ecl_linear_algebra
-    ecl_threads::ecl_threads
-)
-
-set_target_properties(${PROJECT_NAME}
-  PROPERTIES
-    SOVERSION ${${PROJECT_NAME}_VERSION}
-    VERSION ${${PROJECT_NAME}_VERSION}
-)
-
-install(TARGETS ${PROJECT_NAME} EXPORT ${PROJECT_NAME}
-   RUNTIME DESTINATION bin
-   ARCHIVE DESTINATION lib
-   LIBRARY DESTINATION lib
-)
-
